+++ jobbergate-api/CHANGELOG.md
@@ -5,6 +5,7 @@ This file keeps track of all notable changes to jobbergate-api
 ## Unreleased
 
 - Enabled template and job-script files to be renamed on upsert routes [PENG-2070]
+- Added a response model for PUT on `/job-scripts/{id}/upload/{file_type}`
 
 ## 5.2.0a0 -- 2024-04-29